Steps 28-29: add window size to global state
Due to the book using a global variable, and our reproduction of global
variables using lazy_static!, and borrowing rules, reading the window
size moves into the lazy initializer for the global EditorConfig. I
can't find a way of passing mutable refs to its fields to
get_window_size, which means responding to window resizes will be tough.
But also, this whole thing needs to be refactored to be a bit more OOP
with no global variables at some point, so whatever.